requirements.txt
numpy
scipy
pandas
fasttext
mecab-python3
gensim
sklearn
